+++ CHANGELOG.md
@@ -14,7 +14,7 @@ This project adheres to Semantic Versioning.
 * `Rectangle::combine` has been added, which can be used to get a rectangle that contains two other rectangles.
 * Fonts can now have their filter mode set, instead of always using nearest neighbor filtering. ([@tesselode](https://github.com/tesselode) in [#246](https://github.com/17cupsofcoffee/tetra/pull/246)) 
     * Fonts now also respect the default filter mode when they are created.
-
+* Text now supports word-wrapping, via the `Text::wrapped` constructor or the `Text::set_max_width` method.
 
 ### Fixed